+++ cmd/avatar_chat_universal/tty_unix.go
@@ -1,0 +1,22 @@
+//go:build !windows
+
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"golang.org/x/term"
+)
+
+func setupRawTTY() (restore func(), err error) {
+	fd := int(os.Stdin.Fd())
+	if !term.IsTerminal(fd) {
+		return func() {}, nil
+	}
+	state, err := term.MakeRaw(fd)
+	if err != nil {
+		return nil, fmt.Errorf("make tty raw: %w", err)
+	}
+	return func() { _ = term.Restore(fd, state) }, nil
+}